aws-snowflake-kinesis-airflow-pipeline/requirements.txt
apache-airflow==2.7.1
apache-airflow-providers-snowflake==4.1.0
apache-airflow-providers-amazon==7.3.0
boto3==1.28.38
snowflake-connector-python==3.0.4
snowflake-sqlalchemy==1.4.7
pyyaml==6.0.1
pandas==2.0.3 